+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## unreleased
 
+* Deprecate `get_dashboard_by_name` on Grafana 8 and higher, as it no longer
+  supports getting dashboards by slug. Thanks, @oz123.
+
 
 ## 3.6.0 (2023-07-30)